// client/src/redux/comment/comment.types.ts
export enum CommentTypes {
  FETCH_COMMENTS_START = "FETCH_COMMENTS_START",
  FETCH_COMMENTS_SUCCESS = "FETCH_COMMENTS_SUCCESS",
  FETCH_COMMENTS_FAILURE = "FETCH_COMMENTS_FAILURE",
  ADD_COMMENT_START = "ADD_COMMENT_START",
  ADD_COMMENT_SUCCESS = "ADD_COMMENT_SUCCESS",
  ADD_COMMENT_FAILURE = "ADD_COMMENT_FAILURE",
  DELETE_COMMENT_START = "DELETE_COMMENT_START",
  DELETE_COMMENT_SUCCESS = "DELETE_COMMENT_SUCCESS",
  DELETE_COMMENT_FAILURE = "DELETE_COMMENT_FAILURE",
  LIKE_COMMENT_START = "LIKE_COMMENT_START",
  LIKE_COMMENT_SUCCESS = "LIKE_COMMENT_SUCCESS",
  LIKE_COMMENT_FAILURE = "LIKE_COMMENT_FAILURE",
  TOGGLE_COMMENTS_VISIBILITY = "TOGGLE_COMMENTS_VISIBILITY",
}